Move updatable saved query types to /models
