+++ src/client.spec.ts
@@ -29,7 +29,6 @@ import {
 	SystemStatusCategory,
 	SystemStatusMessageReceived,
 	SystemStatusMessageSent,
-	UpdatableSavedQuery,
 	UpdatableScheduledQuery,
 	UpdatableScheduledScript,
 	UpdatableTemplate,
@@ -88,6 +87,7 @@ import {
 	UpdatableNotification,
 	UpdatablePlaybook,
 	UpdatableResource,
+	UpdatableSavedQuery,
 	UploadableAutoExtractor,
 	User,
 	UserPreferences,
