+++ src/models/saved-query/index.ts
@@ -10,6 +10,9 @@ export * from './creatable-saved-query';
 export * from './is-saved-query';
 export * from './raw-creatable-saved-query';
 export * from './raw-saved-query';
+export * from './raw-updatable-saved-query';
 export * from './saved-query';
 export * from './to-raw-creatable-saved-query';
+export * from './to-raw-updatable-saved-query';
 export * from './to-saved-query';
+export * from './updatable-saved-query';
